+++ test/integration/fixture/set1/hr/person/preConditions/prio1PreconditionAllCommands.js
@@ -1,0 +1,9 @@
+module.exports = require('../../../../../../../').definePreCondition({
+  name: '',
+  description: 'Fails if firstname is rumpelstilz',
+  priority: 1
+}, function (command) {
+  if (command.payload.firstname === 'rumpelstilz') {
+    throw new Error('Precondition with prio 1 failed (all commands)');
+  }
+});